chore: move balance types in its own file
